fix: invalidate fork menu after live agent refresh
